implement get_matching_text() to return characters found and confidence (#5)
* implement get_matching_text() to return characters found and confidence

get_matching_text() uses openCV methods to open an image, grayscale it, binarize it,
and pad it. Then, a rotation angle is calculated using openCV minAreaRect().
After rotating the image by the negative of this angle, we can assume that
the text within the image is either correctly oriented, rotated by 90 degrees,
rotated by -90 degrees, or upside down. Thus, we must check each of these four
cases for an occurrence of a letter, returning the letter we find in each along with
its confidence.